Fix mocha tests and refactor
Moved server unit tests to `test/server/`. Connected server unit tests to test db.
*IMPORTANT:* use `npm run test-mocha` to run server unit tests or it will not find them.
