--- Makefile
@@ -1,15 +1,0 @@
-REPORTER = spec
-
-test:
-	@NODE_ENV=test ./node_modules/.bin/mocha \
-		--reporter $(REPORTER) \
-		--ui tdd
-
-test-w:
-	@NODE_ENV=test ./node_modules/.bin/mocha \
-		--reporter $(REPORTER) \
-		--growl \
-		--ui tdd \
-		--watch
-
-.PHONY: test test-w
