--- server/public/index.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-</head>
-
-<body>
-  woo
-</body>
-
-</html>
